Fix code bugs exposed by strict compile on FreeBSD 10.1

+++ src/FSAL/common_pnfs.c
@@ -145,12 +145,14 @@ nfsstat4 FSAL_encode_ipv4_netaddr(XDR *xdrs, uint16_t proto, uint32_t addr,
 
 	/* Then we convert the address and port to a string and encode it. */
 
-	written_length =
-	    snprintf(addrbuff, v4_addrbuff_len, "%hhu.%hhu.%hhu.%hhu.%hhu.%hhu",
-		     ((addr & 0xff000000) >> 0x18),
-		     ((addr & 0x00ff0000) >> 0x10),
-		     ((addr & 0x0000ff00) >> 0x08), (addr & 0x000000ff),
-		     ((port & 0xff00) >> 0x08), (port & 0x00ff));
+	written_length = snprintf(addrbuff, v4_addrbuff_len,
+				  "%u.%u.%u.%u.%u.%u",
+				  (unsigned int) ((addr & 0xff000000) >> 0x18),
+				  (unsigned int) ((addr & 0x00ff0000) >> 0x10),
+				  (unsigned int) ((addr & 0x0000ff00) >> 0x08),
+				  (unsigned int) (addr & 0x000000ff),
+				  (unsigned int) ((port & 0xff00) >> 0x08),
+				  (unsigned int) (port & 0x00ff));
 	if (written_length >= v4_addrbuff_len) {
 		LogCrit(COMPONENT_FSAL,
 			"Programming error in FSAL_encode_ipv4_netaddr defined in %s:%u %s causing snprintf to overflow address buffer.",
